Introduce operations tree as a new `optree` module (#110)
An operations tree (optree) will be used as *an alternative backend
representation for a program* instead of a syntax tree (ast). The latter
is relatively easy to build (since it is bijective with a program source
code), but hard to analyze. An optree, conversely, is intended to be
walked through and used for a program analysis, but it is also can be
built easily from a syntax tree.

Each node of an optree represents an abstract operation that comes from
a language domain. *Operation* may have multiple operands (input
values), results (output values), and nested operations of any kind
inside its body. *Value*, along with a parent operation, holds its data
type and a list of uses (to memoize which operations has the given value
as an operand).

It is important to notice that values are simple SSA registers, which
means that being defined once, it cannot be redefined, and this fact
makes it easy to track definition-use chain per each value. This concept
is widely used in such frameworks like LLVM IR, MLIR.